+++ Helper/Reviews.php
@@ -2,8 +2,11 @@
 
 namespace Feedaty\Badge\Helper;
 
+use Feedaty\Badge\Helper\ConfigRules;
 use Magento\Framework\App\Helper\AbstractHelper;
 use Magento\Store\Api\StoreRepositoryInterface;
+use Magento\Review\Model\ResourceModel\Review\CollectionFactory;
+use Magento\Review\Model\ResourceModel\Rating\CollectionFactory as RatingCollectionFactory;
 
 
 class Reviews extends AbstractHelper
@@ -33,6 +36,18 @@ class Reviews extends AbstractHelper
      */
     protected $_orderRepository;
 
+    /**
+     * @var \Feedaty\Badge\Helper\ConfigSetting
+     */
+    protected $_helperConfigRules;
+    /**
+     * @var CollectionFactory
+     */
+    private $reviewCollection;
+    /**
+     * @var RatingCollectionFactory
+     */
+    private $ratingCollectionFactory;
 
     /**
      * @param \Psr\Log\LoggerInterface $logger
@@ -46,7 +61,11 @@ class Reviews extends AbstractHelper
         \Magento\Review\Model\ReviewFactory           $reviewFactory,
         \Magento\Review\Model\RatingFactory           $ratingFactory,
         StoreRepositoryInterface                      $storeRepository,
-        \Magento\Sales\Api\OrderRepositoryInterface   $orderRepository
+        \Magento\Sales\Api\OrderRepositoryInterface   $orderRepository,
+        ConfigRules $helperConfigRules,
+        CollectionFactory       $reviewCollection,
+        RatingCollectionFactory $ratingCollectionFactory
+
     )
     {
         $this->_logger = $logger;
@@ -54,6 +73,9 @@ class Reviews extends AbstractHelper
         $this->_ratingFactory = $ratingFactory;
         $this->_storeRepository = $storeRepository;
         $this->_orderRepository = $orderRepository;
+        $this->_helperConfigRules = $helperConfigRules;
+        $this->reviewCollection = $reviewCollection;
+        $this->ratingCollectionFactory = $ratingCollectionFactory;
     }
 
     /**
@@ -66,7 +88,7 @@ class Reviews extends AbstractHelper
         try {
             $review = $this->_reviewFactory->create()->load($reviewData['review_id']);
         } catch (\Exception $e) {
-            $this->_logger->info("Feedaty error disableReview :  " . $e->getMessage());
+            $this->_logger->info("Feedaty | Error can not disable Review :  " . $e->getMessage());
         }
 
         if (!is_null($review)) {
@@ -84,7 +106,7 @@ class Reviews extends AbstractHelper
         try {
             $review = $this->_reviewFactory->create()->load($reviewData['review_id']);
         } catch (\Exception $e) {
-            $this->_logger->info("Feedaty Error mediateReview :  " . $e->getMessage());
+            $this->_logger->info("Feedaty | Error cannot mediate Review :  " . $e->getMessage());
         }
 
         if (!is_null($review)) {
@@ -100,21 +122,29 @@ class Reviews extends AbstractHelper
      */
     public function getStoreViewIdByOrder($orderId)
     {
+        $forceDefaultStore = $this->_helperConfigRules->getReviewForceDefaultStore();
 
-        $order = null;
-        try {
-            $order = $this->_orderRepository->get($orderId);
+        if($forceDefaultStore === "1"){
+            $defaultStore = $this->_helperConfigRules->getReviewDefaultStore();
+            return $defaultStore;
         }
-        catch (\Exception $e) {
-            $this->_logger->info("Feedaty Error : order id does not exist " . $orderId . " Error message". $e->getMessage());
+        else{
+            $order = null;
+            try {
+                $order = $this->_orderRepository->get($orderId);
+            }
+            catch (\Exception $e) {
+                $this->_logger->info("Feedaty | Error : order id does not exist " . $orderId . " Error message". $e->getMessage());
+            }
+
+            if (!is_null($order)) {
+                $websiteId = $order->getStore()->getWebsiteId();
+                return $websiteId;
+            }
+
+            return null;
         }
 
-        if (!is_null($order)) {
-            $websiteId = $order->getStore()->getWebsiteId();
-            return $websiteId;
-        }
-
-        return null;
 
     }
 
@@ -141,7 +171,8 @@ class Reviews extends AbstractHelper
      */
     public function getReviewCollection($productId, $feedatyId)
     {
-        $collection = $this->_reviewFactory->create()->getCollection()
+
+        $collection = $this->reviewCollection->create()
             ->addEntityFilter(
                 'product',
                 $productId
@@ -153,6 +184,7 @@ class Reviews extends AbstractHelper
             ->setDateOrder();
 
         return $collection->getData();
+
     }
 
     /*
@@ -160,7 +192,7 @@ class Reviews extends AbstractHelper
      */
     public function getAllFeedatyReviewCount()
     {
-        $collection = $this->_reviewFactory->create()->getCollection()
+        $collection = $this->reviewCollection->create()
             ->addFieldToFilter(
                 'feedaty_source',
                 1
@@ -179,7 +211,7 @@ class Reviews extends AbstractHelper
      */
     public function getAllFeedatyRemovedReviewCount()
     {
-        $collection = $this->_reviewFactory->create()->getCollection()
+        $collection = $this->reviewCollection->create()
             ->addStatusFilter(
                 \Magento\Review\Model\Review::STATUS_NOT_APPROVED)
             ->addFieldToFilter(
@@ -200,7 +232,7 @@ class Reviews extends AbstractHelper
      */
     public function getAllFeedatyMediatedReviewCount()
     {
-        $collection = $this->_reviewFactory->create()->getCollection()
+        $collection = $this->reviewCollection->create()
             ->addFieldToFilter(
                 'feedaty_product_mediated',
                 1
@@ -219,7 +251,7 @@ class Reviews extends AbstractHelper
      */
     public function getLastFeedatyReviewCreated()
     {
-        $collection = $this->_reviewFactory->create()->getCollection()
+        $collection = $this->reviewCollection->create()
             ->addFieldToFilter(
                 'feedaty_source',
                 1
@@ -237,15 +269,15 @@ class Reviews extends AbstractHelper
 
 
     /**
-     * @param $feedatyid
+     * @param $feedatyId
      * @return mixed
      */
-    public function getAllReviewsByFeedatyId($feedatyid)
+    public function getAllReviewsByFeedatyId($feedatyId)
     {
-        $collection = $this->_reviewFactory->create()->getCollection()
+        $collection = $this->reviewCollection->create()
             ->addFieldToFilter(
                 'feedaty_product_review_id',
-                $feedatyid
+                $feedatyId
             )
             ->setOrder(
                 'review_id',
@@ -257,11 +289,14 @@ class Reviews extends AbstractHelper
     }
 
     /**
-     * @return mixed
-     * @throws \Magento\Framework\Exception\LocalizedException
+     * @return array|null
      */
-    public function getRatingCollection(){
-        $ratingCollection = $this->_ratingFactory->create()->getResourceCollection()->addEntityFilter('product')->load();
+    public function getRatingCollection()
+    {
+        $ratingCollection = $this->ratingCollectionFactory->create()
+            ->addEntityFilter('product')
+            ->load();
+
         return $ratingCollection->getData();
     }